添加 material风格的设置界面 Libary 添加Trina的若干工具类 SettingActivity  Bug

+++ app/src/main/java/com/cyt/ieasy/widget/MyGridLayout.java
@@ -24,7 +24,7 @@ public class MyGridLayout extends ViewGroup {
     if (attrs != null) {
       TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.MyGridLayout);
       colums = a.getInteger(R.styleable.MyGridLayout_numColumns, 2);
-      margin = (int) a.getInteger(R.styleable.MyGridLayout_itemMargin, 2);
+      margin = a.getInteger(R.styleable.MyGridLayout_itemMargin, 2);
     }
   }
